Phase 5: render flutter_scene on web through the shim
flutter_scene now renders a real 3D scene (spinning cuboid, UnlitMaterial)
on web via the WebGL2 shim. The smoke test's new Scene tab drives the full
flutter_scene pipeline - scene graph, render graph (scene pass -> tonemap
pass), depth-tested geometry - end to end.

flutter_scene + flutter_scene_importer:
  * Swapped all flutter_gpu imports to flutter_gpu_shim (+ deps). On native
    the shim re-exports flutter_gpu, so this is a no-op there.
  * Shader loading is now async: loadBaseShaderLibrary() runs inside
    Scene.initializeStaticResources(); the sync baseShaderLibrary getter
    keeps a native lazy fallback but on web must be preloaded (web asset
    reads can't be synchronous). flutter_scene_importer gains a web
    platform entry.

Shim changes needed to host flutter_scene's real workload:
  * Texture.asImage() is now synchronous on web (blit to the OffscreenCanvas,
    transferToImageBitmap, createImageFromImageBitmap - both sync on
    CanvasKit and Skwasm), so flutter_scene's synchronous
    render(camera, canvas) -> drawImageRect path works unchanged.
  * DeviceBuffer keeps separate element vs non-element GL buffers backed by
    one staging array. flutter_scene packs vertices and indices into a
    single DeviceBuffer and binds sub-ranges as both; WebGL2 forbids one
    buffer object being both, so the shim splits them transparently.
  * Implemented the depth / cull / winding / blend state setters (were
    Phase-3 stubs). setDepthCompareOperation toggles DEPTH_TEST + depthFunc;
    depthMask is forced on before depth clears. This fixes face ordering.
  * Corrected Viewport to match flutter_gpu (int fields + DepthRange) and
    added DepthRange / Scissor; depthClearValue default aligned to 0.0.

Known follow-ups (tracked in docs): verify the tonemap flip_y orientation
on web; PBR/StandardFragment path (IBL prefilter, mat3 normal matrices,
sampler arrays) not yet exercised; full example app (model loading) not
yet ported.

+++ examples/flutter_gpu_shim_smoke/lib/hdr_tab.dart
@@ -124,14 +124,7 @@ class _HdrTabState extends State<HdrTab> {
         ),
         3,
       );
-      pass.setViewport(
-        gpu.Viewport(
-          x: 0,
-          y: 0,
-          width: _size.toDouble(),
-          height: _size.toDouble(),
-        ),
-      );
+      pass.setViewport(gpu.Viewport(x: 0, y: 0, width: _size, height: _size));
       pass.draw();
       pass.clearBindings();
       cmd.submit();
